// packages/client/src/store/affiliate/actionTypes.ts
export enum AffiliateActionTypes {
  SetAffiliateAccount = 'AFFILIATE.SET_ACCOUNT',
  GetAffiliateAccount = 'AFFILIATE.GET_ACCOUNT',
  GetAffiliateCampaigns = 'AFFILIATE.GET_CAMPAIGNS',
  SetAffiliateCampaigns = 'AFFILIATE.SET_CAMPAIGNS',
  CreateAffiliateCampaigns = 'AFFILIATE.CREATE_CAMPAIGNS',
  GetAffiliateReferrals = 'AFFILIATE.GET_REFERRALS',
  SetAffiliateReferrals = 'AFFILIATE.SET_REFERRALS',
  GetAffiliateReferralCode = 'AFFILIATE.GET_REFERRAL_CODE',
  SetAffiliateReferralCode = 'AFFILIATE.SET_REFERRAL_CODE',
  AffiliateWithdrawTransfer = 'AFFILIATE.WITHDRAW_TRANSFER',
}